refactor store, use persist for AsyncStorage
