+++ components/LoadingIndicator/index.ts
@@ -1,0 +1,1 @@
+export { default as LoadingIndicator } from './LoadingIndicator';
