+++ components/SearchBar/SearchBar.tsx
@@ -1,39 +1,13 @@
 
-import React, { useEffect } from 'react';
+import React from 'react';
 import { withTheme, DefaultTheme } from 'styled-components/native';
-import { ActivityIndicator, TouchableOpacity } from 'react-native';
 
 import { SearchView } from './styles';
-import { Icon} from '../Styled';
-import { StopSearch } from '../StopSearch/'
-
-import { useDeparturesStore } from '../../store';
+import { StopSearch } from '../StopSearch'
+import { LoadingIndicator } from '../LoadingIndicator';
 
 interface SearchBarProps {
   theme: DefaultTheme;
-}
-
-interface LoadingIndicatorProps {
-  theme: DefaultTheme;
-}
-
-function LoadingIndicator ({ theme }: LoadingIndicatorProps) {
-  const {
-     isLoading,
-    fetchDepartures,
-  } = useDeparturesStore();
-
-  return (
-    isLoading
-      ? <ActivityIndicator color={theme.colors.text} />
-      : (
-        <TouchableOpacity
-          onPress={() => fetchDepartures()}
-        >
-          <Icon name="reload" size={20} />
-        </TouchableOpacity>
-      )
-  )
 }
 
 const SearchBar = ({ theme }: SearchBarProps) => {
@@ -41,7 +15,7 @@ const SearchBar = ({ theme }: SearchBarProps) => {
   return (
     <SearchView>
       <StopSearch />
-      <LoadingIndicator theme={theme} />
+      <LoadingIndicator />
     </SearchView>
   )
 };
