+++ components/index.ts
@@ -3,3 +3,4 @@ export { default as SearchBar } from './SearchBar/SearchBar';
 export { default as StopSearch } from './StopSearch/StopSearch';
 export { default as TextInput } from './TextInput/TextInput';
 export { default as DepartureBoard } from './DepartureBoard/DepartureBoard';
+export { default as LoadingIndicator } from './LoadingIndicator/LoadingIndicator';
